refactor : 피드백반영2 (1.미사용코드 삭제, 2.Name클래스 생성(원시값 포장)
