+++ src/main/java/racing/Main.java
@@ -1,16 +1,15 @@
 package racing;
 
 import racing.controller.RaceGame;
-import racing.util.NameUtil;
 import racing.view.InputView;
 
 public class Main {
     public static void main(String[] args) {
 
-        String carNames = InputView.insertNamesOfCars();
+        String[] carNames = InputView.insertNamesOfCars();
         int attemptsNum = InputView.insertNumberOfAttempts();
 
         RaceGame raceGame = new RaceGame();
-        raceGame.startRacing(NameUtil.split(carNames), attemptsNum);
+        raceGame.startRacing(carNames, attemptsNum);
     }
 }
