+++ src/main/java/racing/model/Car.java
@@ -5,18 +5,18 @@ import racing.util.RandomUtil;
 
 public class Car implements Comparable<Car> {
 
-    private String name;
+    private Name name;
     private Position position;
     private MoveStrategy moveStrategy;
 
     public Car(String name, MoveStrategy moveStrategy) {
-        this.name = name;
+        this.name = new Name(name);
         this.position = new Position();
         this.moveStrategy = moveStrategy;
     }
 
     public Car(String name, int position, MoveStrategy moveStrategy) {
-        this.name = name;
+        this.name = new Name(name);
         this.position = new Position(position);
         this.moveStrategy = moveStrategy;
     }
@@ -32,7 +32,7 @@ public class Car implements Comparable<Car> {
     }
 
     public String getName() {
-        return name;
+        return name.getName();
     }
 
     public boolean isSamePosition(int position) {
